Split AddCommand into AddTaskCommand, AddModuleCommand
